+++ .test_calcipy_project/tests/test_zz_test_project.py
@@ -1,15 +1,11 @@
 """Final test alphabetically (zz) to catch general integration cases."""
 
-from pathlib import Path
-
-import tomli
+import toml
 from test_project import __version__
 
 
 def test_version():
     """Check that PyProject and __version__ are equivalent."""
-    data = Path('pyproject.toml').read_text()
-
-    result = tomli.loads(data)['tool']['poetry']['version']
+    result = toml.load('pyproject.toml')['tool']['poetry']['version']
 
     assert result == __version__