+++ calcipy/__init__.py
@@ -2,7 +2,7 @@
 
 from loguru import logger
 
-__version__ = '0.12.1'
+__version__ = '0.12.2'
 __pkg_name__ = 'calcipy'
 
 logger.disable(__pkg_name__)