+++ calcipy/code_tag_collector.py
@@ -99,13 +99,14 @@ def _search_files(paths_source: Sequence[Path], regex_compiled: Pattern[str]) ->
         except UnicodeDecodeError as err:
             logger.debug('Could not parse: {path_source}', path_source=path_source, err=err)
 
-        if comments := _search_lines(lines, regex_compiled):
+        comments = _search_lines(lines, regex_compiled)
+        if comments:
             matches.append(_Tags(path_source, comments))
 
     return matches
 
 
-@lru_cache
+@lru_cache(maxsize=128)
 @beartype
 def _git_info(cwd: Path) -> Tuple[Path, str]:
     """Collect information about the local git repository.
@@ -247,5 +248,5 @@ def write_code_tag_file(
 
     if report:
         path_tag_summary.write_text(f'{header}\n\n{report}\n\n<!-- {SKIP_PHRASE} -->\n')
-    else:
-        path_tag_summary.unlink(missing_ok=True)
+    elif path_tag_summary.is_file():
+        path_tag_summary.unlink()